fixing api

+++ api/controllers/postController.js
@@ -1,6 +1,9 @@
 const postService = require('../services/postService')
+const validatorController = require('./validatorController')
 
 const getAllPost = function(req, res) {
+    res.header("Access-Control-Allow-Origin", "*");    
+    res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
     return postService.getAllPost()
         .then((data) => {
             res.send(data);
@@ -13,6 +16,11 @@ const getAllPost = function(req, res) {
 }
 
 const createPost = function(req, res){
+    res.header("Access-Control-Allow-Origin", "*");    
+    res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
+    if (!validatorController.validateToken(req)) {
+        return "false";
+    }
     return postService.createPost(req.body, res)
     .then(data => {
         res.send(data)
@@ -23,6 +31,11 @@ const createPost = function(req, res){
 }
 
 const updatePost = function(req, res){
+    res.header("Access-Control-Allow-Origin", "*");    
+    res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
+    if (!validatorController.validateToken(req)) {
+        return "false";
+    }
     return postService.updatePost(req.body, res)
     .then(data => {
         res.send(data)
@@ -33,6 +46,11 @@ const updatePost = function(req, res){
 }
 
 const deletePost = function(req,res){
+    res.header("Access-Control-Allow-Origin", "*");    
+    res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
+    if (!validatorController.validateToken(req)) {
+        return "false";
+    }
     return postService.deletePost(req.body, res)
     .then(data => {
         postService.deleteComment(req.body, res)
@@ -44,12 +62,16 @@ const deletePost = function(req,res){
 }
 
 const searchPostByTitle = function(req, res) {
+    res.header("Access-Control-Allow-Origin", "*");    
+    res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
     return postService.searchPostByTitle(req.body.forumName)
     .then((data) => res.send(data))
     .catch((err) => res.send(err)); 
 }
 
 const getPostByTag = function(req, res){
+    res.header("Access-Control-Allow-Origin", "*");    
+    res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
     return postService.getPostByTag(req.body, res)
     .then(data => {
         res.send(data)
